requirements.txt
uvicorn~=0.27.0
fastapi~=0.109.0
pydantic~=2.5.3
jinja2~=3.1.3
python-arango~=7.9.1